+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## 2.0.0 - 2022-07-03
+- Rewrite to use HTML datasets
+- Theme selector now uses a dialog element
+
 ## 1.0.0 - 2022-04-21
 - Cleaned code
 - Better API